Update closure-templates to v20180314"

+++ java/src/com/google/template/soy/msgs/restricted/SoyMsgPluralPart.java
@@ -19,13 +19,9 @@ package com.google.template.soy.msgs.restricted;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.collect.ImmutableList;
-import com.google.template.soy.msgs.restricted.SoyMsgPart.Case;
-
 import com.ibm.icu.text.PluralRules;
 import com.ibm.icu.util.ULocale;
-
 import java.util.Objects;
-
 import javax.annotation.Nullable;
 
 /**
@@ -43,7 +39,6 @@ public final class SoyMsgPluralPart extends SoyMsgPart {
   /** The various cases for this plural statement. The default statement has a null key. */
   private final ImmutableList<Case<SoyMsgPluralCaseSpec>> cases;
 
-
   /**
    * @param pluralVarName The plural variable name.
    * @param offset The offset for this plural statement.
@@ -57,18 +52,15 @@ public final class SoyMsgPluralPart extends SoyMsgPart {
     this.cases = ImmutableList.copyOf(cases);
   }
 
-
   /** Returns the plural variable name. */
   public String getPluralVarName() {
     return pluralVarName;
   }
 
-
   /** Returns the offset. */
   public int getOffset() {
     return offset;
   }
-
 
   /** Returns the cases. */
   public ImmutableList<Case<SoyMsgPluralCaseSpec>> getCases() {
@@ -79,11 +71,11 @@ public final class SoyMsgPluralPart extends SoyMsgPart {
    * Returns the list of parts to implement the case.
    *
    * @param pluralValue The current plural value
-   * @param locale The locale for interpreting non-specific plural parts.  Allowed to be null if it
-   *     is known that there are no non-specific plural parts (This is commonly the case for
-   *     default messages, since soy only allows direct specification of explicit or 'other').
+   * @param locale The locale for interpreting non-specific plural parts. Allowed to be null if it
+   *     is known that there are no non-specific plural parts (This is commonly the case for default
+   *     messages, since soy only allows direct specification of explicit or 'other').
    */
-  public ImmutableList<SoyMsgPart> lookupCase(int pluralValue, @Nullable ULocale locale) {
+  public ImmutableList<SoyMsgPart> lookupCase(long pluralValue, @Nullable ULocale locale) {
     // TODO(lukes): clean up this method, the control flow is overly complex.  It could be cleaned
     // up by separating cases into 3 different lists.
 
@@ -108,16 +100,13 @@ public final class SoyMsgPluralPart extends SoyMsgPart {
 
       } else {
         hasNonExplicitCases = true;
-
       }
     }
 
     if (caseParts == null && hasNonExplicitCases) {
       // Didn't match any numeric value.  Check which plural rule it matches.
       String pluralKeyword = PluralRules.forLocale(locale).select(pluralValue - offset);
-      SoyMsgPluralCaseSpec.Type correctCaseType =
-          new SoyMsgPluralCaseSpec(pluralKeyword).getType();
-
+      SoyMsgPluralCaseSpec.Type correctCaseType = new SoyMsgPluralCaseSpec(pluralKeyword).getType();
 
       // Iterate the cases once again for non-numeric keywords.
       for (Case<SoyMsgPluralCaseSpec> case0 : getCases()) {
@@ -137,7 +126,8 @@ public final class SoyMsgPluralPart extends SoyMsgPart {
     return checkNotNull(caseParts);
   }
 
-  @Override public boolean equals(Object other) {
+  @Override
+  public boolean equals(Object other) {
     if (!(other instanceof SoyMsgPluralPart)) {
       return false;
     }
@@ -147,8 +137,8 @@ public final class SoyMsgPluralPart extends SoyMsgPart {
         && cases.equals(otherPlural.cases);
   }
 
-
-  @Override public int hashCode() {
+  @Override
+  public int hashCode() {
     return Objects.hash(SoyMsgPluralPart.class, offset, pluralVarName, cases);
   }
 }
